Fix windowSize prop passing issue in component hierarchy
## 🐛 Bug Fix
- **Issue**: TypeError: undefined is not an object (evaluating 'windowSize.width')
- **Cause**: Layout component was trying to clone React elements that couldn't receive props
- **Solution**: Move windowSize state to ForecastViz and pass directly to DataVisualization

## 🔧 Changes Made
- **ForecastViz**: Added windowSize state and useEffect for window resize handling
- **Layout**: Simplified to only handle sidebar, removed window size management
- **DataVisualization**: Removed unnecessary windowSize props for RSV and NHSN views
- **Build**: Confirmed successful build with no errors

## ✅ Result
- Fixed runtime error when loading visualizations
- Cleaner component prop passing
- Maintained functionality while fixing architecture issue
